Refactoring on inferred tuple names (VB)
