+++ src/Compilers/CSharp/Portable/Syntax/SyntaxFacts.cs
@@ -431,9 +431,9 @@ namespace Microsoft.CodeAnalysis.CSharp
         /// Given an initializer expression infer the name of anonymous property or tuple element.
         /// Returns null if unsuccessful
         /// </summary>
-        public static string TryGetInferredMemberName(this ExpressionSyntax input)
+        public static string TryGetInferredMemberName(this ExpressionSyntax expression)
         {
-            var nameToken = input.ExtractAnonymousTypeMemberName();
+            var nameToken = expression.ExtractAnonymousTypeMemberName();
             return nameToken.Kind() == SyntaxKind.IdentifierToken ? nameToken.ValueText : null;
         }
