+++ src/Compilers/VisualBasic/Portable/Syntax/SyntaxFacts.vb
@@ -1,5 +1,6 @@
 ﻿' Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
+Imports Microsoft.CodeAnalysis.VisualBasic.Symbols
 Imports Microsoft.CodeAnalysis.VisualBasic.Syntax
 ' NOTE: The other partial class definition is generated by a tool.
 ' Because VB doesn't support Partial Module definitions the actual SyntaxFacts implementations are in this class
@@ -1170,6 +1171,18 @@ Namespace Microsoft.CodeAnalysis.VisualBasic
             End While
 
             Return False
+        End Function
+
+        ''' <summary>
+        ''' Checks whether the element name Is reserved.
+        '''
+        ''' For example:
+        ''' "Item3" Is reserved (at certain positions).
+        ''' "Rest", "ToString" And other members of System.ValueTuple are reserved (in any position).
+        ''' Names that are Not reserved return false.
+        ''' </summary>
+        Public Shared Function IsReservedTupleElementName(elementName As String) As Boolean
+            Return TupleTypeSymbol.IsElementNameReserved(elementName) <> -1
         End Function
     End Class
 End Namespace
