+++ src/Features/CSharp/Portable/IntroduceVariable/CSharpIntroduceVariableService_IntroduceLocal.cs
@@ -166,51 +166,6 @@ namespace Microsoft.CodeAnalysis.CSharp.IntroduceVariable
                 && !typeSymbol.IsFormattableString();
         }
 
-        private static async Task<Tuple<SemanticDocument, ISet<ExpressionSyntax>>> ComplexifyParentingStatements(
-            SemanticDocument semanticDocument,
-            ISet<ExpressionSyntax> matches,
-            CancellationToken cancellationToken)
-        {
-            // First, track the matches so that we can get back to them later.
-            var newRoot = semanticDocument.Root.TrackNodes(matches);
-            var newDocument = semanticDocument.Document.WithSyntaxRoot(newRoot);
-            var newSemanticDocument = await SemanticDocument.CreateAsync(newDocument, cancellationToken).ConfigureAwait(false);
-            var newMatches = newSemanticDocument.Root.GetCurrentNodes(matches.AsEnumerable()).ToSet();
-
-            // Next, expand the topmost parenting expression of each match, being careful
-            // not to expand the matches themselves.
-            var topMostExpressions = newMatches
-                .Select(m => m.AncestorsAndSelf().OfType<ExpressionSyntax>().Last())
-                .Distinct();
-
-            newRoot = await newSemanticDocument.Root
-                .ReplaceNodesAsync(
-                    topMostExpressions,
-                    computeReplacementAsync: async (oldNode, newNode, ct) =>
-                    {
-                        return await Simplifier
-                            .ExpandAsync(
-                                oldNode,
-                                newSemanticDocument.Document,
-                                expandInsideNode: node =>
-                                {
-                                    var expression = node as ExpressionSyntax;
-                                    return expression == null
-                                        || !newMatches.Contains(expression);
-                                },
-                                cancellationToken: ct)
-                            .ConfigureAwait(false);
-                    },
-                    cancellationToken: cancellationToken)
-                .ConfigureAwait(false);
-
-            newDocument = newSemanticDocument.Document.WithSyntaxRoot(newRoot);
-            newSemanticDocument = await SemanticDocument.CreateAsync(newDocument, cancellationToken).ConfigureAwait(false);
-            newMatches = newSemanticDocument.Root.GetCurrentNodes(matches.AsEnumerable()).ToSet();
-
-            return Tuple.Create(newSemanticDocument, newMatches);
-        }
-
         private Document RewriteExpressionBodiedMemberAndIntroduceLocalDeclaration(
             SemanticDocument document,
             ArrowExpressionClauseSyntax arrowExpression,
