+++ src/Features/VisualBasic/Portable/IntroduceVariable/VisualBasicIntroduceVariableService.Rewriter.vb
@@ -34,7 +34,7 @@ Namespace Microsoft.CodeAnalysis.VisualBasic.IntroduceVariable
                 Dim newNode = MyBase.VisitParenthesizedExpression(node)
                 If node IsNot newNode AndAlso newNode.IsKind(SyntaxKind.ParenthesizedExpression) Then
                     Dim parenthesizedExpression = DirectCast(newNode, ParenthesizedExpressionSyntax)
-                    Dim innerExpression = DirectCast(parenthesizedExpression.OpenParenToken.GetNextToken().Parent, ExpressionSyntax)
+                    Dim innerExpression = parenthesizedExpression.OpenParenToken.GetNextToken().Parent
                     If innerExpression.HasAnnotation(_replacementAnnotation) AndAlso innerExpression.Equals(parenthesizedExpression.Expression) Then
                         Return newNode.WithAdditionalAnnotations(Simplifier.Annotation)
                     End If
